+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added implementation for `Brep.from_breps` in `compas_rhino.geometry.RhinoBrep`.
 * Added implementation for `Brep.from_torus` in `compas_rhino.geometry.RhinoBrep`.
 * Added implementation for `Brep.from_polygons` in `compas_rhino.geometry.RhinoBrep`.
+* Added implementation for `Brep.from_pipe` in `compas_rhino.geometry.RhinoBrep`.
 
 ### Changed